+++ cli/src/main/java/org/jboss/as/cli/CommandContext.java
@@ -96,22 +96,6 @@ public interface CommandContext {
         printLine(message);
     }
 
-    default String readLine(String prompt, boolean password) throws CommandLineException {
-        try {
-            return input(prompt, password);
-        } catch (InterruptedException ex) {
-            throw new CommandLineException(ex);
-        }
-    }
-
-    default String input(String prompt, boolean password) throws CommandLineException, InterruptedException {
-        return null;
-    }
-
-    default int[] input() throws CommandLineException, InterruptedException {
-        return null;
-    }
-
     /**
      * Prints a collection of strings as columns to the CLI's output.
      * @param col  the collection of strings to print as columns.
