+++ cli/src/main/java/org/jboss/as/cli/impl/aesh/CLICommandContainer.java
@@ -27,7 +27,6 @@ import org.aesh.command.impl.parser.CommandLineCompletionParser;
 import org.aesh.command.impl.parser.CommandLineParser;
 import org.aesh.command.populator.CommandPopulator;
 import org.aesh.command.validator.OptionValidatorException;
-import org.aesh.console.AeshContext;
 import org.aesh.command.invocation.InvocationProviders;
 import org.aesh.command.Command;
 import org.aesh.command.CommandException;
@@ -41,6 +40,7 @@ import org.aesh.command.validator.CommandValidatorException;
 import org.aesh.complete.AeshCompleteOperation;
 import org.aesh.parser.ParsedLineIterator;
 import org.aesh.parser.ParsedLine;
+import org.aesh.readline.AeshContext;
 import org.jboss.as.cli.CommandContext;
 import org.jboss.as.cli.impl.aesh.commands.operation.LegacyCommandContainer;
 import org.wildfly.core.cli.command.aesh.CLICommandInvocation;
