+++ cli/src/main/java/org/wildfly/core/cli/command/aesh/CLICompleterInvocation.java
@@ -23,10 +23,10 @@ package org.wildfly.core.cli.command.aesh;
 
 import java.util.Collection;
 import java.util.List;
-import org.aesh.console.AeshContext;
 import org.aesh.command.Command;
 import org.aesh.command.completer.CompleterInvocation;
 import org.aesh.command.registry.CommandRegistry;
+import org.aesh.readline.AeshContext;
 import org.aesh.readline.terminal.formatting.TerminalString;
 import org.jboss.as.cli.CommandContext;
