+++ core-feature-pack/src/main/resources/modules/system/layers/base/org/jboss/as/cli/main/module.xml
@@ -68,12 +68,14 @@
         <module name="javax.api"/>
         <module name="ibm.jdk" />
         <module name="sun.jdk" />
-        
+
+        <!--
         <system>
             <paths>
                 <path name="META-INF/services"/>
             </paths>
         </system>
+        -->
     </dependencies>
 
 </module>
