+++ pom.xml
@@ -102,7 +102,7 @@
         <version.org.codehaus.woodstox.woodstox-core>5.0.3</version.org.codehaus.woodstox.woodstox-core>
         <version.org.fusesource.jansi>1.16</version.org.fusesource.jansi>
         <version.org.glassfish.javax.json>1.0.4</version.org.glassfish.javax.json>
-        <version.org.aesh-readline>1.1-SNAPSHOT</version.org.aesh-readline>
+        <version.org.aesh-readline>1.1</version.org.aesh-readline>
         <version.org.aesh>1.0-SNAPSHOT</version.org.aesh>
         <version.org.jboss.byteman>3.0.6</version.org.jboss.byteman>
         <version.org.jboss.classfilewriter>1.2.1.Final</version.org.jboss.classfilewriter>
